flask3/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
gradio_client==0.16.2
pillow==10.3.0
python-multipart==0.0.9
